+++ gorethink.go
@@ -30,9 +30,10 @@ func SetVerbose(verbose bool) {
 	Log.Level = logrus.InfoLevel
 }
 
-// SupportJSONTag allows you to enable the use of the JSON tag. When true is
-// passed to this function GoRethink will start checking for the JSON tag, the
-// GoRethink will still be used and the driver will check for this tag first.
-func SupportJSONTag(v bool) {
-	encoding.SupportJSONTag = v
+// SetTags allows you to override the tags used when decoding or encoding
+// structs. The driver will check for the tags in the same order that they were
+// passed into this function. If nil is passed then the driver will just check
+// for the gorethink tag.
+func SetTags(tags ...string) {
+	encoding.Tags = tags
 }